Reorganize controller classes and interfaces
