+++ weave/Menus/ControllerJoinTest.tscn
@@ -1,4 +1,4 @@
-[gd_scene load_steps=2 format=3 uid="uid://blodracpvvq4t"]
+[gd_scene load_steps=2 format=3 uid="uid://suha7hqfnwef"]
 
 [ext_resource type="Script" path="res://Scripts/MenuControllers/ControllerJoinTest.cs" id="1_t5hqe"]
