+++ weave/Scripts/Controller/Controller.cs
@@ -1,0 +1,7 @@
+namespace weave.Controller;
+
+public enum Controller
+{
+    Keyboard,
+    Gamepad
+}
