+++ weave/Scripts/Controller/IController.cs
@@ -1,8 +1,9 @@
-namespace weave.InputHandlers;
+namespace weave.Controller;
 
 public interface IController
 {
     bool IsTurningLeft();
     bool IsTurningRight();
     int DeviceId => -1;
+    public Controller Type { get; }
 }
