+++ weave/Scripts/Controller/KeyboardController.cs
@@ -1,19 +1,11 @@
-using System.Collections.Generic;
 using Godot;
 
-namespace weave.InputHandlers;
+namespace weave.Controller;
 
 public sealed class KeyboardController : IController
 {
     private readonly Key _left;
     private readonly Key _right;
-    public static readonly IReadOnlyList<(Key, Key)> Keybindings = new List<(Key, Key)>
-    {
-        (Key.Left, Key.Right),
-        (Key.Key1, Key.Q),
-        (Key.B, Key.N),
-        (Key.Z, Key.X)
-    };
 
     public KeyboardController((Key, Key) keybindings)
     {
@@ -30,4 +22,6 @@ public sealed class KeyboardController : IController
     {
         return Input.IsKeyPressed(_right);
     }
+
+    public Controller Type => Controller.Keyboard;
 }
